refactor: rename `command_runner` to `exchanger`
To add an exchanger of transfer TRBs.
